requirements.txt
edge-tts
mutagen
pypdf2
python-dotenv
EbookLib
beautifulsoup4